requirements.txt
Appium-Python-Client
pytest
selenium
pytest-html
numpy
pytest-rerunfailures